feat: Change the list_id type and comment_id type to long

+++ src/main/java/todo/entity/ToDoList.java
@@ -18,7 +18,7 @@ public class ToDoList {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "list_id", nullable = false)
-    private int listId;
+    private long listId;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "email", nullable = false)
